measure attendance freshness by school
Our original thought was that attendance would get blocked up at the SIS/District-level, but in practice there is a lot of variance by school, with some schools falling quite a bit behind the rest of the district.

Breaking this down to the school level should make this more useful for diagnostics. This is still compatible with the downstream test, which is agnostic about the columns in this table.