draft of optionally not requiring stu_ed_org to reach dim_student
Concept: for some downstream reporting purposes, it is confusing for a
record not to be in, e.g. 'enrollment' because the student is not in
stuEdOrg -- this is an implicit requirement that is hard for folks to
understand. We may decide to remove this logic and _always_ left join in
the future, but this branch puts it behind an optional flag to let us
try it out before committing.